fix: move `types.d.ts` into bridge package directory
